moved most of other tests to their folders
