--- asset/src/match/index.ts
@@ -1,7 +1,0 @@
-import Processor from './processor';
-import Schema from '../transform/schema';
-
-export {
-    Processor,
-    Schema
-};
