+++ asset/src/match/schema.ts
@@ -1,0 +1,3 @@
+import TransformSchema from '../transform/schema';
+
+export default TransformSchema;
