+++ test/match/processor-spec.ts
@@ -1,23 +1,20 @@
-import 'jest-extended';
 import path from 'path';
 import { OpTestHarness } from 'teraslice-test-harness';
 import { DataEntity, newTestExecutionConfig } from '@terascope/job-components';
-import { Processor, Schema } from '../asset/src/selection';
-import { makeTest } from './helpers';
+import { Processor, Schema } from '../asset/src/match';
+import { makeTest } from '../helpers';
 
-describe('selection phase', () => {
+describe('match phase', () => {
     const testAssetPath = path.join(__dirname, './assets');
-    let opTest: OpTestHarness;
     const type = 'processor';
+    let opTest: OpTestHarness;
     const assetName = 'someAssetId';
+
     const opConfig = {
-        _op: 'transform',
-        rules: [`${assetName}:transformRules.txt`],
+        _op: 'watcher',
         plugins: ['someAssetId:plugins'],
-        type_config: { location: 'geo-point' },
-        variables: {
-            foo: 'data'
-        }
+        rules: [`${assetName}:matchRules.txt`],
+        types: { _created: 'date' }
     };
 
     const executionConfig = newTestExecutionConfig({
@@ -33,17 +30,17 @@ describe('selection phase', () => {
 
     afterAll(() => opTest.shutdown());
 
-    it('can run and select data', async () => {
+    it('can return matching documents', async () => {
         const data = DataEntity.makeArray([
-            { some: 'data', isTall: true },
-            { some: 'thing else', person: {} },
-            { hostname: 'www.example.com', bytes: 1000 },
-            { location: '33.435967,-111.867710', zip: 94302 },
-            {}
+            { some: 'data', bytes: 1200 },
+            { some: 'data', bytes: 200 },
+            { some: 'other', bytes: 1200 },
+            { other: 'xabcd' },
+            { _created: '2018-12-16T15:16:09.076Z' }
         ]);
 
         const results = await opTest.run(data);
-        // because of the * selector
-        expect(results).toBeArrayOfSize(5);
+
+        expect(results.length).toEqual(3);
     });
 });
