+++ test/selection/processor-spec.ts
@@ -1,22 +1,27 @@
+import 'jest-extended';
 import path from 'path';
 import { OpTestHarness } from 'teraslice-test-harness';
 import { DataEntity, newTestExecutionConfig } from '@terascope/job-components';
-import { Processor, Schema } from '../asset/src/output';
-import { makeTest } from './helpers';
+import { Processor, Schema } from '../../asset/src/selection';
+import { makeTest } from '../helpers';
 
-describe('output phase', () => {
+describe('selection phase', () => {
     const testAssetPath = path.join(__dirname, './assets');
     let opTest: OpTestHarness;
     const type = 'processor';
-
+    const assetName = 'someAssetId';
     const opConfig = {
         _op: 'transform',
+        rules: [`${assetName}:transformRules.txt`],
         plugins: ['someAssetId:plugins'],
-        rules: ['someAssetId:transformRules.txt']
+        type_config: { location: 'geo-point' },
+        variables: {
+            foo: 'data'
+        }
     };
 
     const executionConfig = newTestExecutionConfig({
-        assets: ['someAssetId'],
+        assets: [assetName],
         operations: [opConfig]
     });
 
@@ -28,15 +33,17 @@ describe('output phase', () => {
 
     afterAll(() => opTest.shutdown());
 
-    it('can run and validate data', async () => {
-        const data = [
-            new DataEntity({ interm1: 'hello', interm2: 'world', final: 'hello world' }),
-            new DataEntity({ lastField: 'someValue' }),
-        ];
+    it('can run and select data', async () => {
+        const data = DataEntity.makeArray([
+            { some: 'data', isTall: true },
+            { some: 'thing else', person: {} },
+            { hostname: 'www.example.com', bytes: 1000 },
+            { location: '33.435967,-111.867710', zip: 94302 },
+            {}
+        ]);
 
         const results = await opTest.run(data);
-
-        expect(results.length).toEqual(1);
-        expect(results[0]).toEqual({ final: 'hello world' });
+        // because of the * selector
+        expect(results).toBeArrayOfSize(5);
     });
 });
