+++ test/set_key/processor-spec.ts
@@ -1,15 +1,15 @@
 import 'jest-extended';
 import { DataEntity, cloneDeep, OpConfig } from '@terascope/job-components';
-import DataWindow from '../asset/src/__lib/data-window';
-import Processor from '../asset/src/remove_key/processor';
-import Schema from '../asset/src/remove_key/schema';
-import { makeTest } from './helpers';
+import DataWindow from '../../asset/src/__lib/data-window';
+import Processor from '../../asset/src/set_key/processor';
+import Schema from '../../asset/src/set_key/schema';
+import { makeTest } from '../helpers';
 
 const opConfig: OpConfig = {
     _op: 'set_key',
     field: 'name'
 };
-
+// TODO: remove lodash from package.json
 const testData = [
     {
         id: 1,
@@ -31,6 +31,7 @@ describe('set_key should', () => {
     beforeAll(async () => {
         await testHarness.initialize({ opConfig, type: 'processor' });
     });
+
     afterAll(() => testHarness.shutdown());
 
     it('generate an empty result if no input data', async () => {
@@ -38,24 +39,24 @@ describe('set_key should', () => {
         expect(results).toBeArrayOfSize(0);
     });
 
-    it('return docs as data entities with no field as the key', async () => {
+    it('return docs as data entities with name field as the key', async () => {
         const results = await testHarness.run(testData) as DataEntity[];
 
         results.forEach((doc) => expect(DataEntity.isDataEntity(doc)).toBe(true));
-
-        expect(results[0].getMetadata('_key')).toBe(undefined);
-        expect(results[1].getMetadata('_key')).toBe(undefined);
-        expect(results[2].getMetadata('_key')).toBe(undefined);
+        expect(results[0].getMetadata('_key')).toBe('joe');
+        expect(results[1].getMetadata('_key')).toBe('moe');
+        expect(results[2].getMetadata('_key')).toBe('randy');
     });
 
     it('return data entities with the name field as the key', async () => {
-        const newTestData = cloneDeep(testData).map((doc: any) => DataEntity.make(doc, { _key: 'key' }));
+        const newTestData = cloneDeep(testData).map((doc: any) => DataEntity.make(doc, { _key: 'id' }));
+
         const results = await testHarness.run(newTestData) as DataEntity[];
 
         results.forEach((doc) => expect(DataEntity.isDataEntity(doc)).toBe(true));
-        expect(results[0].getMetadata('_key')).toBe(undefined);
-        expect(results[1].getMetadata('_key')).toBe(undefined);
-        expect(results[2].getMetadata('_key')).toBe(undefined);
+        expect(results[0].getMetadata('_key')).toBe('joe');
+        expect(results[1].getMetadata('_key')).toBe('moe');
+        expect(results[2].getMetadata('_key')).toBe('randy');
     });
 
     it('return data window with data entities metadata _key field as the key', async () => {
@@ -67,11 +68,11 @@ describe('set_key should', () => {
         const results = await testHarness.run(testWindow) as DataEntity[];
 
         results.forEach((doc) => expect(DataEntity.isDataEntity(doc)).toBe(true));
-        expect(results[0].asArray()[0].getMetadata('_key')).toBe(undefined);
-        expect(results[0].asArray()[1].getMetadata('_key')).toBe(undefined);
-        expect(results[0].asArray()[2].getMetadata('_key')).toBe(undefined);
-        expect(results[1].asArray()[0].getMetadata('_key')).toBe(undefined);
-        expect(results[1].asArray()[1].getMetadata('_key')).toBe(undefined);
-        expect(results[1].asArray()[2].getMetadata('_key')).toBe(undefined);
+        expect(results[0].asArray()[0].getMetadata('_key')).toBe('joe');
+        expect(results[0].asArray()[1].getMetadata('_key')).toBe('moe');
+        expect(results[0].asArray()[2].getMetadata('_key')).toBe('randy');
+        expect(results[1].asArray()[0].getMetadata('_key')).toBe('floe');
+        expect(results[1].asArray()[1].getMetadata('_key')).toBe('noe');
+        expect(results[1].asArray()[2].getMetadata('_key')).toBe('blandy');
     });
 });
