+++ test/sort/processor-spec.ts
@@ -1,9 +1,9 @@
 import 'jest-extended';
 import { DataEntity, OpConfig } from '@terascope/job-components';
-import Processor from '../asset/src/sort/processor';
-import Schema from '../asset/src/sort/schema';
-import DataWindow from '../asset/src/__lib/data-window';
-import { makeTest } from './helpers';
+import Processor from '../../asset/src/sort/processor';
+import Schema from '../../asset/src/sort/schema';
+import DataWindow from '../../asset/src/__lib/data-window';
+import { makeTest } from '../helpers';
 
 const testData = [
     {
